refactor(mobile): split unified home screen
