+++ apps/mobile/src/features/home/home.styles.ts
@@ -1,0 +1,818 @@
+import { StyleSheet } from 'react-native'
+import { border, fontSize, lineHeight, palette, radius, size, spacing } from '../../theme'
+import {
+  UNIFIED_CHANNEL_ICON_TILE_SIZE,
+  UNIFIED_CHANNEL_LIST_PADDING,
+  UNIFIED_CHANNEL_ROW_PADDING,
+  UNIFIED_CREATE_MENU_POINTER_SIZE,
+  UNIFIED_CREATE_MENU_WIDTH,
+  UNIFIED_HEADER_LEFT_PADDING,
+  UNIFIED_HOME_ACCENT_COLOR,
+  UNIFIED_HOME_BASE_COLOR,
+  UNIFIED_HOME_BORDER_COLOR,
+  UNIFIED_HOME_SECTION_GAP,
+  UNIFIED_HOME_SURFACE_COLOR,
+  UNIFIED_HOME_SURFACE_MUTED_COLOR,
+  UNIFIED_HOME_TEXT_COLOR,
+  UNIFIED_HOME_TEXT_MUTED_COLOR,
+  UNIFIED_HOME_TEXT_SECONDARY_COLOR,
+} from './constants'
+
+export const styles = StyleSheet.create({
+  container: { flex: 1 },
+  headerCoverOpacityMask: {
+    flex: 1,
+    backgroundColor: 'transparent',
+  },
+
+  unifiedRoot: {
+    flex: 1,
+    flexDirection: 'row',
+  },
+  unifiedPageCoverLayer: {
+    position: 'absolute',
+    top: spacing.none,
+    left: spacing.none,
+    right: spacing.none,
+    overflow: 'hidden',
+    backgroundColor: UNIFIED_HOME_BASE_COLOR,
+  },
+  unifiedPageCoverMask: {
+    ...StyleSheet.absoluteFillObject,
+  },
+  unifiedPageCover: {
+    ...StyleSheet.absoluteFillObject,
+  },
+  unifiedRail: {
+    width: size.plusPanelIcon + spacing.md,
+    borderRightWidth: border.none,
+    overflow: 'hidden',
+    position: 'relative',
+    zIndex: 8,
+    elevation: 8,
+  },
+  unifiedRailBodyFade: {
+    position: 'absolute',
+    left: spacing.none,
+    right: spacing.none,
+    bottom: spacing.none,
+    zIndex: 0,
+  },
+  unifiedRailSafeArea: {
+    flex: 1,
+    alignItems: 'center',
+    gap: spacing.sm,
+    paddingTop: spacing.sm,
+  },
+  unifiedRailScroller: {
+    flex: 1,
+    width: '100%',
+  },
+  unifiedRailList: {
+    alignItems: 'center',
+    gap: spacing.xs,
+    paddingBottom: spacing.xl,
+  },
+  unifiedRailDivider: {
+    width: size.iconButtonSm,
+    height: border.hairline,
+    borderRadius: radius.full,
+    backgroundColor: UNIFIED_HOME_SURFACE_MUTED_COLOR,
+    marginVertical: spacing.xs,
+  },
+  unifiedRailCreateTouch: {
+    width: '100%',
+    minHeight: size.plusPanelIcon + spacing.sm,
+    alignItems: 'center',
+    justifyContent: 'center',
+    zIndex: 12,
+    elevation: 12,
+  },
+  unifiedRailCreateButton: {
+    width: size.plusPanelIcon,
+    height: size.plusPanelIcon,
+    borderRadius: radius['2xl'],
+    borderWidth: StyleSheet.hairlineWidth,
+    alignItems: 'center',
+    justifyContent: 'center',
+    overflow: 'hidden',
+    shadowOpacity: 0.18,
+    shadowRadius: 24,
+    shadowOffset: { width: spacing.none, height: spacing.xs },
+    elevation: 12,
+  },
+  unifiedRailServerTouch: {
+    width: size.plusPanelIcon + spacing.sm,
+    height: size.plusPanelIcon + spacing.sm,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  unifiedRailAvatar: {
+    width: size.plusPanelIcon,
+    height: size.plusPanelIcon,
+    borderRadius: radius.xl,
+    borderWidth: border.active,
+    alignItems: 'center',
+    justifyContent: 'center',
+    overflow: 'hidden',
+  },
+  unifiedRailDirectAvatar: {
+    width: size.plusPanelIcon,
+    height: size.plusPanelIcon,
+    borderRadius: radius.full,
+    borderWidth: StyleSheet.hairlineWidth,
+    alignItems: 'center',
+    justifyContent: 'center',
+    overflow: 'hidden',
+  },
+  unifiedRailUnread: {
+    position: 'absolute',
+    right: spacing.xs,
+    bottom: spacing.xs,
+    minWidth: size.badgeMd,
+    height: size.badgeMd,
+    borderRadius: radius.full,
+    alignItems: 'center',
+    justifyContent: 'center',
+    paddingHorizontal: spacing.xs,
+  },
+  unifiedRailBadgeText: {
+    color: palette.white,
+    fontSize: fontSize.micro,
+    fontWeight: '900',
+  },
+  unifiedPanel: {
+    flex: 1,
+    minWidth: 0,
+  },
+  unifiedPager: {
+    flex: 1,
+    position: 'relative',
+    zIndex: 1,
+  },
+  unifiedPagerPage: {
+    flex: 1,
+  },
+  unifiedWorkspacePanel: {
+    flex: 1,
+    borderWidth: StyleSheet.hairlineWidth,
+    borderTopWidth: border.none,
+    borderLeftWidth: border.none,
+    borderBottomWidth: border.none,
+    borderRadius: radius.none,
+    overflow: 'hidden',
+    position: 'relative',
+  },
+  unifiedWorkspaceBodyBackdrop: {
+    position: 'absolute',
+    left: spacing.none,
+    right: spacing.none,
+    bottom: spacing.none,
+    zIndex: 0,
+  },
+  unifiedWorkspaceHeader: {
+    borderBottomWidth: border.none,
+    overflow: 'visible',
+    position: 'relative',
+    zIndex: 3,
+  },
+  unifiedWorkspaceHeaderSafe: {
+    flex: 1,
+  },
+  unifiedHeaderCover: {
+    ...StyleSheet.absoluteFillObject,
+  },
+  unifiedMasthead: {
+    flex: 1,
+    justifyContent: 'flex-start',
+    paddingLeft: UNIFIED_HEADER_LEFT_PADDING,
+    paddingRight: spacing.sm,
+    paddingTop: spacing.tight,
+    paddingBottom: spacing.md,
+  },
+  unifiedHeaderContent: {
+    minHeight: size.controlLg,
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.md,
+    paddingHorizontal: spacing.none,
+    paddingVertical: spacing.sm,
+    overflow: 'visible',
+  },
+  unifiedHeaderAvatarShadow: {
+    borderRadius: radius.xl,
+    shadowColor: palette.black,
+    shadowOpacity: 0.12,
+    shadowRadius: 10,
+    shadowOffset: { width: spacing.none, height: spacing.xs },
+    elevation: 8,
+  },
+  unifiedServerIdentityPressable: {
+    flex: 1,
+    minWidth: 0,
+    minHeight: size.controlLg,
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.md,
+    borderRadius: radius.xl,
+    paddingRight: spacing.sm,
+    overflow: 'visible',
+  },
+  unifiedServerIdentity: {
+    flex: 1,
+    minWidth: 0,
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.md,
+  },
+  unifiedServerTitleBlock: {
+    flex: 1,
+    minWidth: 0,
+    gap: spacing.xxs,
+    paddingVertical: spacing.sm,
+    overflow: 'visible',
+  },
+  unifiedServerTitle: {
+    lineHeight: lineHeight.lg,
+    paddingVertical: spacing.xxs,
+  },
+  unifiedServerTitleOnCover: {
+    textShadowColor: palette.black,
+    textShadowOffset: { width: spacing.none, height: spacing.xxs },
+    textShadowRadius: 6,
+  },
+  unifiedHeaderIconOnCover: {
+    shadowColor: palette.black,
+    shadowOpacity: 0.12,
+    shadowRadius: 2,
+    shadowOffset: { width: spacing.none, height: spacing.xxs },
+  },
+  unifiedHomeText: {
+    color: UNIFIED_HOME_TEXT_COLOR,
+  },
+  unifiedHomeSecondaryText: {
+    color: UNIFIED_HOME_TEXT_SECONDARY_COLOR,
+  },
+  unifiedHomeMutedText: {
+    color: UNIFIED_HOME_TEXT_MUTED_COLOR,
+  },
+  unifiedHomeSubtleText: {
+    color: palette.neutral500,
+  },
+  unifiedSectionHeaderText: {
+    color: UNIFIED_HOME_TEXT_MUTED_COLOR,
+    fontSize: fontSize.xs,
+    fontWeight: '800',
+    lineHeight: lineHeight.xs,
+  },
+  unifiedSectionHeaderRow: {
+    minHeight: size.badgeLg,
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.xs,
+  },
+  unifiedSectionCountTag: {
+    minWidth: size.badgeLg,
+    height: size.badgeLg,
+    borderRadius: radius.full,
+    paddingHorizontal: spacing.xs,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: UNIFIED_HOME_SURFACE_MUTED_COLOR,
+    borderWidth: border.hairline,
+    borderColor: UNIFIED_HOME_BORDER_COLOR,
+  },
+  unifiedSectionCountText: {
+    color: UNIFIED_HOME_TEXT_SECONDARY_COLOR,
+    fontSize: fontSize.micro,
+    fontWeight: '900',
+    lineHeight: lineHeight.xs,
+  },
+  unifiedTouchPressed: {
+    transform: [{ scale: 0.985 }],
+  },
+  unifiedSectionLabel: {
+    fontSize: fontSize.xs,
+    fontWeight: '800',
+    lineHeight: lineHeight.xs,
+  },
+  unifiedChannelList: {
+    paddingLeft: UNIFIED_CHANNEL_LIST_PADDING,
+    paddingRight: spacing.md,
+    paddingTop: spacing.xs,
+    paddingBottom: spacing.lg,
+  },
+  unifiedChannelListHeader: {
+    marginBottom: spacing.md,
+  },
+  unifiedChannelScroll: {
+    flex: 1,
+  },
+  unifiedShortcutStage: {
+    gap: UNIFIED_HOME_SECTION_GAP,
+    marginBottom: UNIFIED_HOME_SECTION_GAP,
+  },
+  unifiedShortcutGroup: {
+    gap: spacing.tight,
+  },
+  unifiedShortcutHeader: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.xs,
+    paddingLeft: UNIFIED_CHANNEL_ROW_PADDING,
+    paddingRight: spacing.none,
+  },
+  unifiedShortcutPager: {
+    gap: spacing.sm,
+    paddingRight: spacing.md,
+  },
+  unifiedShortcutPage: {
+    flexDirection: 'row',
+    alignItems: 'flex-start',
+    gap: spacing.sm,
+  },
+  unifiedAppTrack: {
+    display: 'none',
+    width: size.thumbnailMd,
+    height: size.dotXs,
+    borderRadius: radius.full,
+    marginTop: spacing.none,
+    marginLeft: spacing.xs,
+  },
+  unifiedShortcutTile: {
+    minHeight: size.avatarXl,
+    borderRadius: radius.xl,
+    alignItems: 'center',
+    justifyContent: 'center',
+    gap: spacing.xs,
+    paddingHorizontal: spacing.xs,
+    paddingVertical: spacing.tight,
+  },
+  unifiedShortcutLabel: {
+    maxWidth: '100%',
+    textAlign: 'center',
+  },
+  unifiedServerAppIcon: {
+    width: size.avatarMd,
+    height: size.avatarMd,
+    borderRadius: radius.lg,
+    borderWidth: StyleSheet.hairlineWidth,
+    alignItems: 'center',
+    justifyContent: 'center',
+    overflow: 'hidden',
+  },
+  unifiedServerAppIconImage: {
+    width: '100%',
+    height: '100%',
+  },
+  unifiedShortcutSkeletonIcon: {
+    width: size.avatarMd,
+    height: size.avatarMd,
+    borderRadius: radius.lg,
+  },
+  unifiedShortcutSkeletonLabel: {
+    width: size.plusPanelIcon,
+    height: size.dotLg,
+    borderRadius: radius.full,
+  },
+  unifiedChannelGroup: {
+    gap: spacing.xxs,
+    marginBottom: UNIFIED_HOME_SECTION_GAP,
+  },
+  unifiedChannelSectionHeader: {
+    marginBottom: spacing.xxs,
+  },
+  unifiedChannelSectionHeaderSpaced: {
+    marginTop: spacing.md,
+  },
+  unifiedChannelListItem: {
+    marginBottom: spacing.xxs,
+  },
+  unifiedGroupHeaderRow: {
+    minHeight: size.controlXs,
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    gap: spacing.sm,
+  },
+  unifiedGroupHeader: {
+    flex: 1,
+    minWidth: 0,
+    minHeight: size.controlXs,
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingLeft: spacing.none,
+    paddingRight: spacing.none,
+    paddingTop: spacing.none,
+    position: 'relative',
+  },
+  unifiedGroupChevron: {
+    position: 'absolute',
+    left: -spacing.xs,
+    width: spacing.lg,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  unifiedGroupTitle: {
+    marginLeft: UNIFIED_CHANNEL_ROW_PADDING,
+  },
+  unifiedGroupCreateButton: {
+    width: size.iconButtonSm,
+    height: size.iconButtonSm,
+    borderRadius: radius.full,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  unifiedWechatChannelRow: {
+    minHeight: size.controlSm,
+    borderWidth: StyleSheet.hairlineWidth,
+    borderRadius: radius.lg,
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.sm,
+    paddingLeft: UNIFIED_CHANNEL_ROW_PADDING,
+    paddingVertical: spacing.xs,
+    paddingRight: spacing.sm,
+  },
+  unifiedChannelIconTile: {
+    width: UNIFIED_CHANNEL_ICON_TILE_SIZE,
+    height: UNIFIED_CHANNEL_ICON_TILE_SIZE,
+    borderRadius: radius.lg,
+    borderWidth: StyleSheet.hairlineWidth,
+    backgroundColor: UNIFIED_HOME_SURFACE_MUTED_COLOR,
+    alignItems: 'center',
+    justifyContent: 'center',
+    overflow: 'visible',
+    position: 'relative',
+  },
+  unifiedWechatChannelBody: {
+    flex: 1,
+    minWidth: 0,
+  },
+  unifiedWechatChannelTitleRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.xs,
+  },
+  unifiedWechatChannelTitle: {
+    flex: 1,
+    minWidth: 0,
+    fontSize: fontSize.sm,
+    fontWeight: '700',
+    lineHeight: lineHeight.sm,
+  },
+  unifiedWechatChannelTitleUnread: {
+    fontWeight: '900',
+  },
+  unifiedChannelUnreadDotInline: {
+    width: size.dotMd,
+    height: size.dotMd,
+    borderRadius: radius.full,
+  },
+  unifiedSkeletonStack: {
+    gap: spacing.sm,
+  },
+  unifiedSkeletonRow: {
+    height: size.homeChannelRowHeight,
+    borderRadius: radius.lg,
+  },
+  unifiedEmptyPanel: {
+    minHeight: size.panelStateMinHeight,
+    borderWidth: StyleSheet.hairlineWidth,
+    borderRadius: radius['2xl'],
+    alignItems: 'center',
+    justifyContent: 'center',
+    gap: spacing.md,
+    padding: spacing.lg,
+  },
+  unifiedEmptyText: {
+    textAlign: 'center',
+  },
+  unifiedEmptyWrap: {
+    gap: spacing.lg,
+    paddingTop: spacing['6xl'],
+  },
+  unifiedDmPickerEmpty: {
+    minHeight: size.panelStateMinHeight,
+    alignItems: 'center',
+    justifyContent: 'center',
+    gap: spacing.md,
+    paddingHorizontal: spacing.lg,
+  },
+  unifiedDmPickerEmptyText: {
+    textAlign: 'center',
+  },
+  unifiedCommandSearchInput: {
+    flex: 1,
+    minWidth: 0,
+    height: lineHeight.lg,
+    fontSize: fontSize.lg,
+    fontWeight: '800',
+    lineHeight: lineHeight.lg,
+    paddingHorizontal: spacing.none,
+    paddingVertical: spacing.none,
+    margin: spacing.none,
+  },
+  unifiedCommandSearchInputAndroid: {
+    textAlignVertical: 'center',
+    includeFontPadding: false,
+  },
+  unifiedCommandSearchInputIos: {
+    transform: [{ translateY: -spacing.tight }],
+  },
+  unifiedCommandSearchClose: {
+    width: size.controlLg,
+    height: size.controlLg,
+    borderRadius: radius.full,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  commandModalSearchIconBox: {
+    width: size.controlMd,
+    height: size.controlMd,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  commandModalSearchInputBox: {
+    flex: 1,
+    minWidth: 0,
+    height: size.plusPanelIconLg,
+    justifyContent: 'center',
+  },
+  unifiedCommandResults: {
+    paddingHorizontal: spacing.sm,
+    paddingVertical: spacing.xs,
+    gap: spacing.none,
+  },
+  unifiedCommandEmptyText: {
+    alignSelf: 'stretch',
+    paddingHorizontal: spacing.xs,
+    paddingVertical: spacing.md,
+    textAlign: 'center',
+  },
+  unifiedPressed: {
+    transform: [{ scale: 0.985 }],
+  },
+  commandModalRoot: {
+    flex: 1,
+    justifyContent: 'flex-end',
+    paddingHorizontal: spacing.sm,
+    paddingBottom: spacing.md,
+  },
+  commandModalResults: {
+    maxHeight: '48%',
+    borderRadius: radius['3xl'],
+    borderWidth: StyleSheet.hairlineWidth,
+    overflow: 'hidden',
+    marginBottom: spacing.tight,
+    shadowOpacity: 0.36,
+    shadowRadius: 24,
+    shadowOffset: { width: spacing.none, height: spacing.sm },
+    elevation: 18,
+  },
+  commandModalSearchPill: {
+    height: size.plusPanelIconLg,
+    borderRadius: radius['3xl'],
+    borderWidth: StyleSheet.hairlineWidth,
+    overflow: 'hidden',
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.xs,
+    paddingHorizontal: spacing.sm,
+    shadowOpacity: 0.48,
+    shadowRadius: 18,
+    shadowOffset: { width: spacing.none, height: spacing.sm },
+    elevation: 20,
+  },
+  createMenuPopover: {
+    position: 'absolute',
+    width: UNIFIED_CREATE_MENU_WIDTH,
+    overflow: 'visible',
+    shadowOpacity: 0.14,
+    shadowRadius: 18,
+    shadowOffset: { width: spacing.none, height: spacing.xs },
+    elevation: 24,
+  },
+  createMenuPanel: {
+    width: '100%',
+    borderRadius: radius['2xl'],
+    borderWidth: StyleSheet.hairlineWidth,
+    paddingHorizontal: spacing.xs,
+    paddingVertical: spacing.tight,
+    overflow: 'hidden',
+    zIndex: 2,
+  },
+  createMenuArrow: {
+    position: 'absolute',
+    top: -UNIFIED_CREATE_MENU_POINTER_SIZE + spacing.xs,
+    width: spacing.none,
+    height: spacing.none,
+    borderLeftWidth: UNIFIED_CREATE_MENU_POINTER_SIZE,
+    borderRightWidth: UNIFIED_CREATE_MENU_POINTER_SIZE,
+    borderBottomWidth: UNIFIED_CREATE_MENU_POINTER_SIZE,
+    borderLeftColor: 'transparent',
+    borderRightColor: 'transparent',
+    zIndex: 3,
+  },
+  createMenuInnerStroke: {
+    ...StyleSheet.absoluteFillObject,
+    borderRadius: radius['2xl'],
+    borderWidth: StyleSheet.hairlineWidth,
+    zIndex: 1,
+  },
+  createMenuBubble: {
+    position: 'relative',
+    zIndex: 2,
+    gap: spacing.none,
+  },
+  createMenuRow: {
+    minHeight: size.controlLg,
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.md,
+    borderRadius: radius.xl,
+    paddingHorizontal: spacing.md,
+    paddingVertical: spacing.xs,
+  },
+  commandModalRow: {
+    minHeight: size.listItemLg,
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.md,
+    borderBottomWidth: StyleSheet.hairlineWidth,
+    paddingHorizontal: spacing.xs,
+    paddingVertical: spacing.sm,
+  },
+  commandModalIcon: {
+    width: size.avatarMd,
+    height: size.avatarMd,
+    borderRadius: radius.lg,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  commandModalTextColumn: {
+    flex: 1,
+    minWidth: 0,
+    gap: spacing.xxs,
+  },
+  commandModalRowLabel: {
+    lineHeight: lineHeight.md,
+  },
+  commandModalRowMeta: {
+    lineHeight: lineHeight.xs,
+  },
+  unifiedSidePage: {
+    flex: 1,
+    paddingHorizontal: spacing.lg,
+    paddingTop: spacing.md,
+    paddingBottom: spacing['3xl'],
+  },
+  unifiedSideTopRow: {
+    minHeight: size.controlMd,
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.sm,
+    paddingBottom: spacing.md,
+  },
+  unifiedPageHeaderText: {
+    flex: 1,
+    minWidth: 0,
+    gap: spacing.xxs,
+  },
+  unifiedSideTitle: {
+    lineHeight: lineHeight.md,
+  },
+  unifiedWorkspaceSectionTitle: {
+    fontSize: fontSize.xs,
+    fontWeight: '800',
+    lineHeight: lineHeight.xs,
+  },
+  unifiedSideIconAction: {
+    width: size.iconButtonMd,
+    height: size.iconButtonMd,
+    borderRadius: radius.full,
+    borderWidth: StyleSheet.hairlineWidth,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: palette.surface,
+  },
+  unifiedStackedIconAction: {
+    width: size.iconButtonLg,
+    height: size.iconButtonLg,
+    borderRadius: radius.full,
+    borderWidth: StyleSheet.hairlineWidth,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: UNIFIED_HOME_SURFACE_COLOR,
+    position: 'relative',
+  },
+  unifiedStackedIconBadge: {
+    position: 'absolute',
+    right: spacing.xxs,
+    bottom: spacing.xxs,
+    width: size.badgeMd,
+    height: size.badgeMd,
+    borderRadius: radius.full,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: UNIFIED_HOME_ACCENT_COLOR,
+    borderWidth: border.hairline,
+    borderColor: UNIFIED_HOME_BASE_COLOR,
+  },
+  unifiedSideListContent: {
+    paddingBottom: spacing['5xl'],
+    gap: spacing.xxs,
+  },
+  unifiedWorkspaceListContent: {
+    gap: spacing.none,
+  },
+  unifiedPreviewRow: {
+    minHeight: size.controlLg,
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.sm,
+    paddingVertical: spacing.xs,
+  },
+  unifiedPreviewRowText: {
+    flex: 1,
+    minWidth: 0,
+    gap: spacing.xxs,
+  },
+  unifiedMemberTreeRow: {
+    minHeight: size.controlLg,
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.sm,
+    paddingVertical: spacing.xs,
+  },
+  unifiedBuddyTreeRow: {
+    paddingLeft: spacing.none,
+  },
+  unifiedMemberTreeGuide: {
+    width: spacing['3xl'],
+    alignSelf: 'stretch',
+    position: 'relative',
+  },
+  unifiedMemberTreeLine: {
+    position: 'absolute',
+    left: size.avatarSm / 2,
+    top: spacing.tight,
+    bottom: spacing.tight,
+    width: StyleSheet.hairlineWidth,
+    borderRadius: radius.full,
+  },
+  unifiedMemberTreeLineLast: {
+    bottom: '50%',
+  },
+  unifiedMemberTreeBranch: {
+    position: 'absolute',
+    left: size.avatarSm / 2,
+    top: '50%',
+    width: spacing['3xl'],
+    height: StyleSheet.hairlineWidth,
+    borderRadius: radius.full,
+  },
+  unifiedMemberNameRow: {
+    minWidth: 0,
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: spacing.xs,
+  },
+  unifiedMemberNameText: {
+    flexShrink: 1,
+  },
+  unifiedFileIcon: {
+    width: size.avatarSm,
+    height: size.avatarSm,
+    borderRadius: radius.lg,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  unifiedSideEmpty: {
+    minHeight: size.controlLg,
+    justifyContent: 'center',
+    paddingHorizontal: spacing.xs,
+  },
+  edgeList: {
+    width: '100%',
+  },
+
+  menuItem: {
+    minHeight: size.listItemLg,
+  },
+  menuLabel: {
+    flex: 1,
+  },
+  input: {
+    height: size.controlLg,
+    borderRadius: radius.full,
+    paddingHorizontal: spacing.lg,
+    fontSize: fontSize.md,
+  },
+  switchRow: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    paddingVertical: spacing.sm,
+  },
+})
